Update README.md
Apologies if I'm reaching here, but the best library I could find that is currently being maintained should be referenced and this lib can be deprecated if that is in fact your ultimate plan. I have notice there hasn't been a lot of activity in the last couple years and was hoping to point visitors of this repo to a great alternative instead of leaving them hanging.